Composite job details, running jobs

+++ src/components/Jobs/components/Job/Job.js
@@ -16,6 +16,7 @@ import {
   withStripes,
   stripesShape,
   IfPermission,
+  useOkapiKy,
 } from '@folio/stripes/core';
 
 import {
@@ -25,18 +26,22 @@ import {
   ConfirmationModal,
   FormattedDate,
   FormattedTime,
+  Layout
 } from '@folio/stripes/components';
 import {
   Progress,
   createOkapiHeaders,
   createUrl,
 } from '@folio/stripes-data-transfer-components';
-import { permissions } from '../../../../utils';
+import { permissions, cancelMultipartJob } from '../../../../utils';
 
 import { jobMetaTypes } from './jobMetaTypes';
 import { jobExecutionPropTypes } from './jobExecutionPropTypes';
 
 import * as API from '../../../../utils/upload';
+import * as CompositeJobFields from '../../../../utils/compositeJobStatus';
+import { trimLeadNumbers } from '../../../../utils/multipartUpload';
+
 import {
   addHrid,
   deleteHrid,
@@ -45,6 +50,82 @@ import {
 } from '../../../../redux/actions/jobExecutionsActionCreator';
 
 import css from './Job.css';
+
+const renderCompositeDetails = (jobEntry, previousProgress = { processed: 0, total: 100 }, updateProgress = noop) => {
+  const {
+    completedSliceAmount,
+    erroredSliceAmount,
+    failedSliceAmount,
+    totalSliceAmount,
+    inProgressRecords,
+    completedRecords,
+    failedRecords,
+  } = CompositeJobFields.collectCompositeJobValues(jobEntry);
+
+  const progress = CompositeJobFields.calculateCompositeProgress(
+    {
+      inProgressRecords,
+      completedRecords,
+      failedRecords
+    },
+    jobEntry.totalRecordsInFile,
+    previousProgress,
+    updateProgress
+  );
+
+  return (
+    <>
+      <FormattedMessage
+        id="ui-data-import.progressRunning"
+        tagName="div"
+      />
+      <Progress
+        current={progress.processed}
+        total={progress.total}
+      />
+      <FormattedMessage
+        id="ui-data-import.jobProgress.partsRemaining"
+        tagName="div"
+        values={{
+          current: totalSliceAmount - (failedSliceAmount + completedSliceAmount),
+          total: totalSliceAmount
+        }}
+      />
+      <FormattedMessage
+        id="ui-data-import.jobProgress.partsProcessed"
+        tagName="div"
+        values={{
+          current: completedSliceAmount,
+          total: totalSliceAmount,
+        }}
+      />
+      <ul className={css.compositeList}>
+        <li className={css.listItem}>
+          <FormattedMessage
+            id="ui-data-import.jobProgress.partsCompleted"
+            tagName="div"
+            values={{ amount: completedSliceAmount }}
+          />
+        </li>
+        <li className={css.listItem}>
+          <FormattedMessage
+            id="ui-data-import.jobProgress.partsCompletedWithErrors"
+            tagName="div"
+            values={{ amount: erroredSliceAmount }}
+          />
+        </li>
+        <li className={css.listItem}>
+          <FormattedMessage
+            id="ui-data-import.jobProgress.partsFailed"
+            tagName="div"
+            values={{ amount: failedSliceAmount }}
+          />
+        </li>
+      </ul>
+    </>
+  );
+};
+
 
 const propTypes = {
   stripes: stripesShape.isRequired,
@@ -63,8 +144,10 @@ const JobComponent = ({
 }) => {
   const [deletionInProgress, setDeletionInProgress] = useState(false);
   const [showCancelJobModal, setShowCancelJobModal] = useState(false);
+  const [compositeProgress, updateCompositeProgress] = useState({ processed: 0, total: 100 });
   const calloutRef = useRef();
   const dispatch = useDispatch();
+  const okapiKy = useOkapiKy();
 
   useEffect(() => {
     return () => dispatch(deleteHrid(job.hrId));
@@ -111,10 +194,14 @@ const JobComponent = ({
     } = job;
 
     try {
-      await API.deleteFile(
-        createJobUrl(id),
-        createOkapiHeaders(okapi),
-      );
+      if (job.compositeDetails) {
+        await cancelMultipartJob(okapiKy, id);
+      } else {
+        await API.deleteFile(
+          createJobUrl(id),
+          createOkapiHeaders(okapi),
+        );
+      }
     } catch (error) {
       setDeletionInProgress(false);
 
@@ -140,6 +227,44 @@ const JobComponent = ({
     hideCancelJobConfirmationModal();
 
     await deleteJob();
+  };
+
+  const renderCancelModalMessage = (jobEntry) => {
+    const {
+      completedSliceAmount,
+      totalSliceAmount,
+    } = CompositeJobFields.collectCompositeJobValues(jobEntry);
+
+    return (
+      <>
+        <FormattedMessage id="ui-data-import.modal.cancelRunningSplitJob.message" />
+        <p>
+          <FormattedMessage
+            id="ui-data-import.modal.cancelRunningSplitJob.message.pleaseNote"
+          />
+        </p>
+        <Layout className={`${css.compositeList} padding-end-gutter padding-start-gutter`} element="ul">
+          <li className={css.listItem}>
+            <FormattedMessage id="ui-data-import.modal.cancelRunningSplitJob.message.noRestart" />
+          </li>
+          <li className={css.listItem}>
+            <FormattedMessage id="ui-data-import.modal.cancelRunningSplitJob.message.noRevert" />
+          </li>
+          <li className={css.listItem}>
+            <FormattedMessage
+              id="ui-data-import.modal.cancelRunningSplitJob.message.jobParts"
+              values={{ current: completedSliceAmount, total: totalSliceAmount }}
+            />
+          </li>
+          <li className={css.listItem}>
+            <FormattedMessage
+              id="ui-data-import.modal.cancelRunningSplitJob.message.remaining"
+              values={{ remaining: totalSliceAmount - completedSliceAmount }}
+            />
+          </li>
+        </Layout>
+      </>
+    );
   };
 
   const {
@@ -169,10 +294,10 @@ const JobComponent = ({
         <span>{name}</span>
         &nbsp;
         <span>
-          {fileName}
+          {trimLeadNumbers(fileName)}
           {isDeletionInProgress && (
             <>
-            &nbsp;
+              &nbsp;
               <FormattedMessage
                 id="ui-data-import.stoppedJob"
                 tagName="span"
@@ -209,7 +334,7 @@ const JobComponent = ({
             {jobMeta && (
               <FormattedMessage
                 id="ui-data-import.recordsCount"
-                values={{ count: total }}
+                values={{ count: job.compositeDetails ? job.totalRecordsInFile : total }}
                 tagName="span"
               />
             )}
@@ -220,7 +345,7 @@ const JobComponent = ({
               />
             </span>
           </div>
-          {jobMeta.showProgress && (
+          {jobMeta.showProgress && !job.compositeDetails && (
             <>
               <FormattedMessage
                 id="ui-data-import.progressRunning"
@@ -232,7 +357,7 @@ const JobComponent = ({
               />
             </>
           )}
-
+          {job.compositeDetails && renderCompositeDetails(job, compositeProgress, updateCompositeProgress)}
           {jobMeta.showPreview && (
             <div className={css.jobPreview}>
               <FormattedMessage id="ui-data-import.readyForPreview" />
@@ -252,12 +377,17 @@ const JobComponent = ({
       <ConfirmationModal
         id="cancel-running-job-modal"
         open={showCancelJobModal}
-        heading={<FormattedMessage id="ui-data-import.modal.cancelRunningJob.header" />}
-        message={
+        heading={job.compositeDetails ?
+          <FormattedMessage id="ui-data-import.modal.cancelRunningSplitJob.header" /> :
+          <FormattedMessage id="ui-data-import.modal.cancelRunningJob.header" />
+          }
+        message={job.compositeDetails ?
+          renderCancelModalMessage(job) :
           <FormattedMessage
             id="ui-data-import.modal.cancelRunningJob.message"
             values={{ break: <br /> }}
-          />}
+          />
+        }
         bodyTag="div"
         confirmLabel={<FormattedMessage id="ui-data-import.modal.cancelRunningJob.confirm" />}
         cancelLabel={<FormattedMessage id="ui-data-import.modal.cancelRunningJob.cancel" />}
